# DOCS/API.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SARVAL API",
    "version": "1.0.0",
    "description": "API del planificador de descargas en silos de SARVAL.\n\nToda la API está protegida con JWT bearer salvo `POST /api/auth/login`. El motor de simulación se recalcula **automáticamente** en cualquier endpoint que cambie el plan (subir Excel, crear/editar/borrar viaje, parada, box, lectura de nivel, franja de productividad o parámetros de tolva). El único recálculo manual con notificación de webhooks es `POST /api/planning/recalculate`.\n\n**Concepto \"plan activo\"**: la mayoría de endpoints aceptan `?plan_id=…`. Si no se pasa, se usa el plan vigente (el de la semana en curso).\n\n**Semana operativa**: Lunes 06:00 → Sábado 22:00 (consumo continuo, también de noche).\n\nVer `DOCS/API.md` en el repo para los flujos típicos."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    },
    {
      "url": "https://sarval.example.com",
      "description": "Producción (ajustar)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SequenceItem": {
        "type": "object",
        "description": "Resultado de la secuenciación (un viaje con su hora final calculada por el motor).",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID del sequence_result"
          },
          "trip_id": {
            "type": "integer"
          },
          "trip_number": {
            "type": "string",
            "example": "5678ABC"
          },
          "day": {
            "type": "string",
            "example": "Martes"
          },
          "hora_prevista": {
            "type": "string",
            "example": "09:30"
          },
          "supplier": {
            "type": "string"
          },
          "producto": {
            "type": "string"
          },
          "tons": {
            "type": "number"
          },
          "critico": {
            "type": "boolean"
          },
          "retraso_h": {
            "type": "number",
            "nullable": true
          },
          "nueva_hora": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "viaje_extra": {
            "type": "boolean"
          },
          "dia_final": {
            "type": "string",
            "description": "Día al que el motor desplazó el viaje",
            "example": "Martes"
          },
          "hora_final": {
            "type": "string",
            "description": "Hora a la que el motor desplazó el viaje",
            "example": "10:00"
          },
          "retraso_capacidad_h": {
            "type": "number",
            "description": "Horas de retraso aplicadas por falta de capacidad",
            "example": 0.5
          },
          "retenido_por_critico": {
            "type": "boolean",
            "description": "true si se retuvo para dar paso a un crítico"
          },
          "tolva_id": {
            "type": "integer"
          },
          "tolva_numero": {
            "type": "integer"
          },
          "tolva_nombre": {
            "type": "string"
          }
        }
      },
      "SimulationStep": {
        "type": "object",
        "description": "Un paso de la simulación del nivel del silo (resolución = paso_minutos de la tolva).",
        "properties": {
          "step_index": {
            "type": "integer",
            "example": 0
          },
          "day": {
            "type": "string",
            "example": "Lunes"
          },
          "time": {
            "type": "string",
            "example": "06:00"
          },
          "entries_tons": {
            "type": "number",
            "description": "Total entradas en el paso (camiones + boxes)",
            "example": 0
          },
          "box_entry_tons": {
            "type": "number",
            "example": 0
          },
          "consumption_tons": {
            "type": "number",
            "example": 6
          },
          "silo_level": {
            "type": "number",
            "example": 14
          },
          "is_stoppage": {
            "type": "boolean",
            "example": false
          },
          "tolva_id": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Petición inválida.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/planning/sequence": {
      "get": {
        "tags": [
          "planning"
        ],
        "summary": "Secuenciación calculada por el motor",
        "parameters": [
          {
            "name": "plan_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Plan a consultar. Si se omite se usa el plan activo (vigente)."
          },
          {
            "name": "tolva_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Filtra por tolva."
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de viajes con su hora final.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SequenceItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/planning/simulation": {
      "get": {
        "tags": [
          "planning"
        ],
        "summary": "Simulación paso a paso del nivel del silo",
        "parameters": [
          {
            "name": "plan_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Plan a consultar. Si se omite se usa el plan activo (vigente)."
          },
          {
            "name": "tolva_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Filtra por tolva."
          }
        ],
        "responses": {
          "200": {
            "description": "Serie de pasos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimulationStep"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Entrar en modo simulación (clona el plan)",
        "description": "Clona el plan indicado (vigente o próxima; por defecto el activo) en un plan aislado con `status = \"simulation\"`. Todos los endpoints normales funcionan sobre el clon pasándole su `plan_id`. Los recálculos del clon NUNCA emiten webhooks/alertas. Máximo una simulación abierta por usuario (409 si ya existe, con su id).",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan_id": {
                    "type": "integer",
                    "description": "Plan a clonar. Por defecto, el plan activo."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Clon creado y recalculado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "simulation_plan_id": {
                      "type": "integer"
                    },
                    "parent_plan_id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "El plan no es la semana vigente ni la próxima."
          },
          "409": {
            "description": "El usuario ya tiene una simulación abierta (se devuelve su `simulation_plan_id`)."
          }
        }
      }
    },
...
    "/api/planning/simulation/{id}/diff": {
      "get": {
        "tags": [
          "planning"
        ],
        "summary": "Resumen de cambios de la simulación vs. el plan real",
        "description": "Altas/bajas/modificados por tabla, KPIs comparados (viajes con retraso, horas paradas, stock mínimo) y `base_changed` si el plan real cambió desde que se clonó. Solo el propietario del clon (o un superadmin).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Id del plan de simulación."
          }
        ],
        "responses": {
          "200": {
            "description": "Diff calculado."
          },
          "403": {
            "description": "La simulación pertenece a otro usuario."
          },
          "404": {
            "description": "Simulación no encontrada."
          }
        }
      }
    },
    "/api/planning/simulation/{id}/apply": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Aplicar la simulación al plan real",
        "description": "En una transacción: sobrescribe los datos del plan real con los del clon (el id del plan real NO cambia) y borra el clon. Después recalcula el plan real con `notify: true` (webhooks/alertas).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Id del plan de simulación."
          }
        ],
        "responses": {
          "200": {
            "description": "Cambios aplicados; incluye el diff aplicado y el resultado del recálculo."
          },
          "403": {
            "description": "La simulación pertenece a otro usuario."
          },
          "404": {
            "description": "Simulación no encontrada."
          }
        }
      }
    },
    "/api/planning/simulation/{id}": {
      "delete": {
        "tags": [
          "planning"
        ],
        "summary": "Cancelar la simulación",
        "description": "Borra el clon y todos sus datos. El plan real no se toca.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Id del plan de simulación."
          }
        ],
        "responses": {
          "200": {
            "description": "Simulación borrada."
          },
          "403": {
            "description": "La simulación pertenece a otro usuario."
          },
          "404": {
            "description": "Simulación no encontrada."
          }
        }
      }
    },
    "/api/planning/recalculate": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Recálculo manual del plan activo",
        "description": "Botón \"Actualizar\" en la UI. Único recálculo con `notify: true` → emite los webhooks configurados. El resto de acciones recalculan automáticamente en silencio.",
        "parameters": [
          {
            "name": "plan_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Plan a consultar. Si se omite se usa el plan activo (vigente)."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "plan_id": {
                      "type": "integer"
                    },
                    "sequence_count": {
                      "type": "integer"
                    },
                    "simulation_count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  }
}